+++ structurizr-export/src/main/java/com/structurizr/export/ilograph/IlographExporter.java
@@ -16,10 +16,41 @@ import java.util.stream.Collectors;
  */
 public class IlographExporter extends AbstractWorkspaceExporter {
 
+    public static final String ILOGRAPH_IMPORTS = "ilograph.imports";
     public static final String ILOGRAPH_ICON = "ilograph.icon";
 
     public WorkspaceExport export(Workspace workspace) {
         IndentingWriter writer = new IndentingWriter();
+
+        // Ilograph imports can be specified in the form:
+        //
+        // AWS:ilograph/aws
+        //
+        // Which gets exported as:
+        //
+        // imports:
+        //  - from: ilograph/aws
+        //    namespace: AWS
+        String commaSeparatedListOfImports = workspace.getProperties().get(ILOGRAPH_IMPORTS);
+        if (!StringUtils.isNullOrEmpty(commaSeparatedListOfImports)) {
+            writer.writeLine("imports:");
+
+            String[] ilographImports = commaSeparatedListOfImports.split(",");
+            for (String ilographImport : ilographImports) {
+                String[] parts = ilographImport.split(":");
+                if (parts.length == 2) {
+                    String namespace = parts[0];
+                    String from = parts[1];
+
+                    writer.writeLine("- from: " + from);
+                    writer.indent();
+                    writer.writeLine("namespace: " + namespace);
+                    writer.outdent();
+                }
+            }
+            writer.writeLine();
+        }
+
         writer.writeLine("resources:");
         writer.indent();