+++ changelog.md
@@ -12,7 +12,8 @@
 - structurizr-dsl: Adds a `!relationships` keyword that can be used to find a set of relationships via an expression.
 - structurizr-dsl: Adds a DSL wrapper around the `structurizr-component` component finder.
 - structurizr-dsl: Adds support for local theme files to be specified via `theme` (https://github.com/structurizr/java/issues/331).
-- structurizr-export: Adds support for icons to the Ilograph exporter.
+- structurizr-export: Adds support for icons to the Ilograph exporter (https://github.com/structurizr/java/issues/332).
+- structurizr-export: Adds support for imports to the Ilograph exporter (https://github.com/structurizr/java/issues/332).
 
 ## 2.2.0 (2nd July 2024)